split makeProgram into separate command files
